# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/projects/{projectId}/prompts/{promptName}": {
      "get": {
        "operationId": "PromptController_findOne",
        "parameters": [
          {
            "name": "promptName",
            "required": true,
            "in": "path",
            "description": "Nombre único del prompt dentro del proyecto",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Prompt encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePromptDto"
                }
              }
            }
          },
          "404": {
            "description": "Proyecto o Prompt no encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtiene un prompt lógico por su nombre dentro de un proyecto",
        "tags": [
          "Prompts"
        ]
      },
      "patch": {
        "operationId": "PromptController_update",
        "parameters": [
          {
            "name": "promptName",
            "required": true,
            "in": "path",
            "description": "Nombre único del prompt a actualizar dentro del proyecto",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePromptDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prompt actualizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePromptDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos."
          },
          "404": {
            "description": "Proyecto, Prompt, Tactic o Tag no encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualiza metadatos de un prompt lógico (descripción, tactic, tags) dentro de un proyecto",
        "tags": [
          "Prompts"
        ]
      },
      "delete": {
        "operationId": "PromptController_remove",
        "parameters": [
          {
            "name": "promptName",
            "required": true,
            "in": "path",
            "description": "Nombre único del prompt a eliminar dentro del proyecto",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Prompt eliminado."
          },
          "404": {
            "description": "Proyecto o Prompt no encontrado."
          },
          "409": {
            "description": "Conflicto al eliminar (revisar relaciones sin Cascade)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Elimina un prompt lógico (y sus versiones asociadas por Cascade) dentro de un proyecto",
        "tags": [
          "Prompts"
        ]
      }
    },
...
    "/api/projects/{projectId}/prompt-assets/{assetKey}": {
      "get": {
        "operationId": "PromptAssetController_findOne",
        "parameters": [
          {
            "name": "assetKey",
            "required": true,
            "in": "path",
            "description": "Key única del asset dentro del proyecto",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Asset encontrado con detalles."
          },
          "404": {
            "description": "Proyecto o Asset no encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtiene un prompt asset por su key dentro de un proyecto",
        "tags": [
          "Prompt Assets"
        ]
      },
      "patch": {
        "operationId": "PromptAssetController_update",
        "parameters": [
          {
            "name": "assetKey",
            "required": true,
            "in": "path",
            "description": "Key única del asset a actualizar",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePromptAssetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Asset actualizado."
          },
          "400": {
            "description": "Datos inválidos."
          },
          "404": {
            "description": "Proyecto o Asset no encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualiza metadatos de un prompt asset (nombre, descripción, etc.) dentro de un proyecto",
        "tags": [
          "Prompt Assets"
        ]
      },
      "delete": {
        "operationId": "PromptAssetController_remove",
        "parameters": [
          {
            "name": "assetKey",
            "required": true,
            "in": "path",
            "description": "Key única del asset a eliminar",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Asset eliminado."
          },
          "404": {
            "description": "Proyecto o Asset no encontrado."
          },
          "409": {
            "description": "Conflicto al eliminar (revisar relaciones sin Cascade)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Elimina un prompt asset (y sus versiones/traducciones por Cascade) dentro de un proyecto",
        "tags": [
          "Prompt Assets"
        ]
      }
    },
    "/api/projects/{projectId}/prompt-versions/{promptVersionId}/links": {
      "post": {
        "operationId": "PromptAssetLinkController_create",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "promptVersionId",
            "required": true,
            "in": "path",
            "description": "Prompt Version ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePromptAssetLinkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Link created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePromptAssetLinkDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden Access to Project."
          },
          "404": {
            "description": "Project, Prompt Version, or Asset Version not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Link an Asset Version to a Prompt Version within a Project",
        "tags": [
          "Prompt Asset Links (within Project/Version)"
        ]
      },
      "get": {
        "operationId": "PromptAssetLinkController_findAll",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "promptVersionId",
            "required": true,
            "in": "path",
            "description": "Prompt Version ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of links.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CreatePromptAssetLinkDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden Access to Project."
          },
          "404": {
            "description": "Project or Prompt Version not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all links for a specific Prompt Version within a Project",
        "tags": [
          "Prompt Asset Links (within Project/Version)"
        ]
      }
    },
    "/api/projects/{projectId}/prompt-versions/{promptVersionId}/links/{linkId}": {
      "get": {
        "operationId": "PromptAssetLinkController_findOne",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "promptVersionId",
            "required": true,
            "in": "path",
            "description": "Prompt Version ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "linkId",
            "required": true,
            "in": "path",
            "description": "Link ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Link found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePromptAssetLinkDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden Access to Project."
          },
          "404": {
            "description": "Project, Prompt Version, or Link not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific link by its ID for a Prompt Version within a Project",
        "tags": [
          "Prompt Asset Links (within Project/Version)"
        ]
      },
      "patch": {
        "operationId": "PromptAssetLinkController_update",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "promptVersionId",
            "required": true,
            "in": "path",
            "description": "Prompt Version ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "linkId",
            "required": true,
            "in": "path",
            "description": "Link ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePromptAssetLinkDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Link updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePromptAssetLinkDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden Access to Project."
          },
          "404": {
            "description": "Project, Prompt Version, or Link not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a specific link by its ID for a Prompt Version within a Project",
        "tags": [
          "Prompt Asset Links (within Project/Version)"
        ]
      },
      "delete": {
        "operationId": "PromptAssetLinkController_remove",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "promptVersionId",
            "required": true,
            "in": "path",
            "description": "Prompt Version ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "linkId",
            "required": true,
            "in": "path",
            "description": "Link ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Link deleted."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden Access to Project."
          },
          "404": {
            "description": "Project, Prompt Version, or Link not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a specific link by its ID for a Prompt Version within a Project",
        "tags": [
          "Prompt Asset Links (within Project/Version)"
        ]
      }
    },
...
  },
  "info": {
    "title": "JAPM API",
    "description": "API para la aplicación de Prompt Engineering JAPM",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "InitialTranslationDto": {
        "type": "object",
        "properties": {
          "languageCode": {
            "type": "string",
            "minLength": 2,
            "maxLength": 10,
            "description": "Código de idioma ISO (e.g., es, en)"
          },
          "promptText": {
            "type": "string",
            "description": "Texto traducido del prompt"
          }
        },
        "required": [
          "languageCode",
          "promptText"
        ]
      },
      "CreatePromptDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "pattern": "/^[a-z0-9_]+$/",
            "description": "Nombre único del prompt (usado como ID)",
            "example": "saludo_bienvenida_cliente"
          },
          "description": {
            "type": "string",
            "description": "Descripción del propósito del prompt.",
            "example": "Prompt inicial para saludar a un cliente."
          },
          "tacticId": {
            "type": "string",
            "description": "ID (nombre) de la táctica conversacional asociada.",
            "example": "small_talk_inicio"
          },
          "tags": {
            "uniqueItems": true,
            "description": "Lista de nombres de etiquetas a asociar.",
            "example": [
              "bienvenida",
              "general"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "promptText": {
            "type": "string",
            "description": "Texto base del prompt para la primera versión (v1.0.0)",
            "example": "Hola {{nombre_cliente}}, bienvenido."
          },
          "initialTranslations": {
            "description": "Traducciones iniciales opcionales para la primera versión",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InitialTranslationDto"
            }
          }
        },
        "required": [
          "name",
          "promptText"
        ]
      },
      "UpdatePromptDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "Nueva descripción del propósito del prompt.",
            "example": "Prompt actualizado para saludar."
          },
          "tacticId": {
            "type": "string",
            "nullable": true,
            "description": "ID de la táctica a asociar, o null para desasociar.",
            "example": "small_talk_continuacion"
          },
          "tagIds": {
            "description": "Lista completa de IDs de etiquetas a asociar (reemplaza las existentes). Array vacío para quitar todas.",
            "example": [
              "cl...uuid1",
              "cl...uuid2"
            ],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "UpdatePromptAssetDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre descriptivo del asset",
            "example": "Saludo Formal (España)"
          },
          "type": {
            "type": "string",
            "description": "Tipo de asset (e.g., texto, json, lista)",
            "example": "texto"
          },
          "description": {
            "type": "string",
            "description": "Descripción más detallada del propósito del asset"
          },
          "category": {
            "type": "string",
            "description": "Categoría para organizar assets (e.g., Saludos, Despedidas)",
            "example": "Saludos"
          },
          "enabled": {
            "type": "boolean",
            "description": "Activa o desactiva el asset"
          },
          "projectId": {
            "type": "string",
            "nullable": true,
            "minLength": 25,
            "maxLength": 25,
            "description": "ID opcional del proyecto al que pertenece el asset (null para desvincular)"
          }
        }
      },
      "CreatePromptAssetLinkDto": {
        "type": "object",
        "properties": {
          "promptVersionId": {
            "type": "string",
            "format": "uuid",
            "description": "ID de la PromptVersion a la que se vincula el asset."
          },
          "assetVersionId": {
            "type": "string",
            "format": "uuid",
            "description": "ID de la PromptAssetVersion que se vincula."
          },
          "usageContext": {
            "type": "string",
            "description": "Contexto de uso o propósito de este asset en el prompt.",
            "example": "Nombre del producto"
          },
          "position": {
            "type": "number",
            "minimum": 0,
            "description": "Orden posicional del asset dentro del prompt (si aplica).",
            "example": 1
          },
          "insertionLogic": {
            "type": "string",
            "description": "Lógica de inserción condicional (formato a definir).",
            "example": "variable_exists:user_name"
          },
          "isRequired": {
            "type": "boolean",
            "description": "Indica si el asset es requerido para el prompt (default: true)."
          }
        },
        "required": [
          "promptVersionId",
          "assetVersionId"
        ]
      },
      "UpdatePromptAssetLinkDto": {
        "type": "object",
        "properties": {
          "usageContext": {
            "type": "string",
            "description": "Contexto de uso o propósito de este asset en el prompt.",
            "example": "Nombre del producto"
          },
          "position": {
            "type": "number",
            "minimum": 0,
            "description": "Orden posicional del asset dentro del prompt (si aplica).",
            "example": 1
          },
          "insertionLogic": {
            "type": "string",
            "description": "Lógica de inserción condicional (formato a definir).",
            "example": "variable_exists:user_name"
          },
          "isRequired": {
            "type": "boolean",
            "description": "Indica si el asset es requerido para el prompt (default: true)."
          }
        }
      },
...
    }
  }
}